requirements.txt
absl-py==2.2.2
accelerate==1.6.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
biopython==1.85
bleach==6.2.0
cachetools==5.5.2
cellpose==4.0.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
Cython==0.29.37
dask==2025.4.1
dask-image==2024.5.3
debugpy==1.8.14
decorator==5.2.1
DeepCell==0.12.10
DeepCell-Tracking==0.6.5
DeepCell_Toolbox==0.12.1
defusedxml==0.7.1
diffusers==0.33.1
distributed==2025.4.1
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.19
fastjsonschema==2.21.1
fastremap==1.16.1
filelock==3.18.0
fill_voids==2.0.8
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
imagecodecs==2025.3.30
imageio==2.37.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nd2==0.10.3
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.2
notebook_shim==0.2.4
numcodecs==0.13.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
ome-types==0.6.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.2.1
PIMS==0.7
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic_core==2.33.2
pydot==1.4.2
Pygments==2.19.1
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
resource-backed-dask-array==0.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roifile==2025.5.10
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.27.0
setproctitle==1.3.6
six==1.17.0
slicerator==1.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spektral==1.0.6
stack-data==0.6.3
sympy==1.14.0
tblib==3.1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.4
tensorflow-addons==0.16.1
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tinycss2==1.4.0
tomli==2.2.1
toolz==1.0.0
torch==2.7.0
torchview==0.2.7
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typeguard==4.4.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xsdata==24.3.1
zarr==2.18.3
zict==3.0.0
zipp==3.21.0
wellmap